third_party/blink/web_tests/virtual/stylable-select-disabled/external/wpt/html/semantics/forms/the-select-element/select-parsing.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] <button>s and <datalist>s should be allowed in <select>.
  assert_equals: expected "\\n  div 1\\n  <button>button</button>\\n  div 2\\n  <datalist>\\n    <option>option</option>\\n  </datalist>\\n  div 3\\n" but got "\\n  div 1\\n  button\\n  div 2\\n  \\n    <option>option</option>\\n  \\n  div 3\\n"
[FAIL] </select> should close <button>.
  assert_equals: expected "\\n  <button>button\\n</button>" but got "\\n  button\\n"
[FAIL] </select> should close <datalist>.
  assert_equals: expected "\\n  <datalist>datalist\\n</datalist>" but got "\\n  datalist\\n"
[FAIL] <select> in <button> in <select> should remove inner <select>.
  assert_equals: expected "\\n  <button>\\n    </button>" but got "\\n  \\n    "
[FAIL] <select> in <select><button><div> should remove inner <select>.
  assert_equals: expected "\\n  <button>\\n    <div>\\n      </div></button>" but got "\\n  \\n    \\n      "
[FAIL] Nested <button>s or <datalist>s in <select> should be dropped.
  assert_equals: expected "\\n<button>\\n</button>\\n\\n<datalist>\\n</datalist>\\n\\n" but got "\\n\\n\\n\\n\\n\\n\\n"
Harness: the test ran to completion.
